% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_datas_shinyCV.R
\name{generate_datas_shinyCV}
\alias{generate_datas_shinyCV}
\title{Create an example dataset to send to feed_shinyCV().}
\usage{
generate_datas_shinyCV()
}
\description{
Create an example dataset to send to feed_shinyCV().
}
\examples{
generate_data_shinyCV()
}
