% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_shinyCV.R
\name{init_shinyCV}
\alias{init_shinyCV}
\title{create a folder for your own CV.}
\usage{
init_shinyCV(cv_path = getwd())
}
\arguments{
\item{cv_path}{The directory where your to create your CV folder}
}
\description{
Copy the shinyCV viewer application to set it up in your desktop folder.
Using publish_shinyCV will allow you to push your CV on shinyapps.io. If you
want to use shiny server instead, copy the files manually.
}
\examples{
init_shinyCV()
}
