% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_shinyCV.R
\name{publish_shinyCV}
\alias{publish_shinyCV}
\title{publish your CV to shinyapps.io.}
\usage{
publish_shinyCV(cv_path)
}
\arguments{
\item{cv_path}{The directory where your CV is stored}
}
\description{
Send your CV to shinyapps.io. You need to have an account before,
and everything properly set up
(see \href{https://shiny.rstudio.com/articles/shinyapps.html}{How to configureshinyapps.io}).
}
\examples{
publish_shinyCV()
}
