% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCV.R
\docType{package}
\name{shinyCV}
\alias{shinyCV}
\alias{shinyCV-package}
\title{shinyCV: Custom CV using AdminLTE2 and Shiny.}
\description{
The shinyCV package provides several tools to build nice CVs
based on the adminLTE2 dashboard template
}
\examples{
\dontrun{
if (interactive()) {
  shinyCV::build_shinyCV()
}
}

}
