% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_shinyCV.R
\name{view_shinyCV}
\alias{view_shinyCV}
\title{Launch the shinyCV web application to view your beautiful CV.
It means that you need to use the build_shinyCV function before.}
\usage{
view_shinyCV()
}
\description{
Launch the shinyCV web application to view your beautiful CV.
It means that you need to use the build_shinyCV function before.
}
\examples{
view_shinyCV()
}
