% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_shinyCV.R
\name{build_shinyCV}
\alias{build_shinyCV}
\title{shinyCV builder.}
\usage{
build_shinyCV(import_datas = NULL)
}
\arguments{
\item{import_datas}{\code{NULL} by default. If not, provide the path where
your CV files are located (see \code{\link{init_shinyCV}}). It should be
yourCV_path/www (do not forget www).}
}
\description{
Launch the shinyCV web application to build your beautiful CV.
}
\examples{
build_shinyCV()
}
