% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_shinyCV.R
\name{feed_shinyCV}
\alias{feed_shinyCV}
\title{Feed a shinyCV with data.
This function passes all the data provided by the user to the cv viewer.
Launch view_shinyCV() to see the results.}
\usage{
feed_shinyCV(profile, about, skills, languages, network, formations, projects,
  tasks, publications, publications_screenshots, talks, courses, internships)
}
\arguments{
\item{profile}{The profile list containing the name, current position
age, interests (a vector of "Computer Sciences", "Biology", "Physiology", "Mathematics",
"Neurosciences", "Web-development"), website url, teaser and image of the user. None of this field
is mandatory. If provided, the image field should contain the path
to access the user profile image.}

\item{about}{The about dataframe containing the phone number (a string "+44 650 36 47"),
mail adress (a string "testatgmail.com"),
location as well as social network profiles, namely, linkedin, twitter,
facebook and github. None of this field is mandatory.}

\item{skills}{The skills dataframe containing the skill name and its value
(from 0 to 100). Each field is mandatory.}

\item{languages}{The language dataframe containing the language name and value
(from 0 to 100). Each field is mandatory.}

\item{network}{The network dataframe containing title, sex ("male" or "female"),
name, mail and phone of the network member.}

\item{formations}{The formations dataframe gathering the users diploma. Fields
are title, topic ("industry", "balance-scale", "database", "eyedropper", "heartbeat",
"music", "money", "paw", "paint-brush", "gamepad"), from (for instance "1900-01-01"),
to (for instance "1900-01-01"),
summary, place, supervisor, grade (number between 0 and 5) and extra
(web link for instance).}

\item{projects}{The projects dataframe containg all projects done by the user.
Fields are title, position, overview, supervisors and place.}

\item{tasks}{The tasks associated to each project (if any). It is a list of dataframes
each containing the following fields: name and status ("wip" or "completed"). The idea
is that for each project, there could be some tasks associated.}

\item{publications}{The publications dataframe containing the reference
(Your name et al., Journal Title, year), abstract and a pubmed_link.}

\item{publications_screenshots}{A list containing the screenshots of all
publications (should be the path to the image).}

\item{talks}{The conferences dataframe containing the title,
from (for instance "1900-01-01"),
to (for instance "1900-01-01"), summary, place, price ("yes" or "no")
as well as website.}

\item{courses}{The courses dataframe containing title, topic, nb_students (numeric),
nb_hours (numeric), from (for instance "1900-01-01"), to (for instance "1900-01-01"),
place, supervisor and syllabus (web link if any).}

\item{internships}{The internships dataframe containing title, topic,
from (for instance "1900-01-01"),
to (for instance "1900-01-01"), place, supervisor,
student_level ("bachelor", "master", "PhD" or "PostDoc") and advert (web link)}
}
\description{
Feed a shinyCV with data.
This function passes all the data provided by the user to the cv viewer.
Launch view_shinyCV() to see the results.
}
\examples{

# load an example dataset
generate_datas_shinyCV()

feed_shinyCV(temp_profile, temp_about, temp_skills, temp_languages, temp_network, temp_formations,
             temp_projects, temp_tasks, temp_publications, publications_screenshots = list(),
             temp_talks, temp_courses, temp_internships)
}
