% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_datas_shinyCV.R
\name{generate_datas_shinyCV}
\alias{generate_datas_shinyCV}
\title{Dataset for shinyCV demo.}
\usage{
generate_datas_shinyCV()
}
\description{
Create an example dataset. If the function \code{\link{view_shinyCV}} is called
with the argument datas equal to NULL and data_source set to manual, this dataset
will be used to show an example.
}
\examples{
generate_data_shinyCV()
}
