% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_shinyCV.R
\name{view_shinyCV}
\alias{view_shinyCV}
\title{Launch the shinyCV web application to view your beautiful CV.}
\usage{
view_shinyCV(cv_mode = "basic", data_source = "manual", datas = NULL)
}
\arguments{
\item{cv_mode}{The content to display in the CV. There are currently 5
flavors: "full" which will display the 5 tabPanels, namely formation,
projects, teaching, publications and talks; "basic" which will display only
formation and projetcs; "academic", which will display formation, projects,
teaching and talks; "academic_light", which will display formation, projects,
and teaching; "research", which will display formations, projects, publications
and talks. By default, cv_mode is set to basic.}

\item{data_source}{If data_source is "manual", it means that datas need to be
generated as in the example of feed_shinyCV(). If data_source is the builder_shinyCV()
("from_cvbuilder") then datas are generated by filling the forms in the dedicated interface
( see build_shinyCV()).}

\item{datas}{If data_source is "manual" and if datas are NULL, an example is generated
using generate_datas_shinyCV; if datas are not NULL, datas should be equal to
the list of datas returned by the feed_shinyCV function. In other words,
datas <- feed_shinyCV(profile, about, skills, languages, network,
                     formations, projects, tasks, publications,
                     publications_screenshots,
                     talks, courses, internships)}
}
\description{
Launch the shinyCV web application to view your beautiful CV.
}
\examples{
view_shinyCV(cv_mode = "basic", data_source = "manual")
}
